another split
- split main.lua extra functions into respective already made files
- updated tile sprites
