+++ main.lua
@@ -9,6 +9,12 @@
 #include src/particle.lua
 #include src/animator.lua
 
+-- Utility/Helper Functions
+#include src/helpers.lua
+
+-- Draw Calls
+#include src/draw_calls.lua
+
 -- Pico8
 function _init()
   reset_game()
@@ -162,413 +168,4 @@ function game_loop()
     enemies_active = false 
     sfx(sfx_data.round_complete)
   end
-end
-
-function spawn_enemy()
-  while enemies_remaining > 0 do 
-    enemy_current_spawn_tick = (enemy_current_spawn_tick + 1) % enemy_required_spawn_ticks
-    if (is_there_something_at(map_data[loaded_map].enemy_spawn_location[1], map_data[loaded_map].enemy_spawn_location[2], enemies)) goto spawn_enemy_continue
-    if (enemy_current_spawn_tick ~= 0) goto spawn_enemy_continue 
-    enemy_data_from_template = increase_enemy_health(enemy_templates[wave_data[wave_round][enemies_remaining]])
-    printh(enemy_data_from_template.hp)
-    add(enemies, Enemy:new(map_data[loaded_map].enemy_spawn_location, enemy_data_from_template))
-    -- else
-    --   local enemy_data_from_template = enemy_templates[wave_data[wave_round][enemies_remaining]]
-    --   add(enemies, Enemy:new(map_data[loaded_map].enemy_spawn_location, enemy_data_from_template))
-    -- end
-    enemies_remaining -= 1
-    ::spawn_enemy_continue:: 
-    yield()
-  end
-end
-
-function kill_enemy(enemy)
-  if (enemy.hp > 0) return
-  coins += enemy.reward
-  del(enemies, enemy)
-end
-
-function destroy_particle(particle)
-  if (not Animator.finished(particle.animator)) return
-  del(particles, particle)
-end
-
-function update_enemy_position(enemy)
-  if (not Enemy.step(enemy)) return
-  enemy.x = pathing[enemy.pos].x
-  enemy.y = pathing[enemy.pos].y
-  enemy.pos += 1
-  if (enemy.pos < #pathing + 1) return
-  player_health -= enemy.damage 
-  del(enemies, enemy)
-end
-
-function parse_path()
-  local path_tiles = {}
-  for iy=0, 16 do
-    for ix=0, 16 do
-      local mx = ix + map_data[loaded_map].mget_shift[1]
-      local my = iy + map_data[loaded_map].mget_shift[2]
-      for _, id in pairs(map_data[loaded_map].path_id) do
-        if (mget(mx, my) == id) then 
-          add(path_tiles, { x = mx, y = my })
-        end
-      end
-    end
-  end
-
-  local path = {}
-  local dir = {x=map_data[loaded_map].movement_direction[1],y=map_data[loaded_map].movement_direction[2]}
-  local ending = {
-    x=map_data[loaded_map].enemy_end_location[1] + map_data[loaded_map].mget_shift[1], 
-    y=map_data[loaded_map].enemy_end_location[2] + map_data[loaded_map].mget_shift[2]
-  }
-  local cur = {
-    x = map_data[loaded_map].enemy_spawn_location[1] + map_data[loaded_map].mget_shift[1] + dir.x, 
-    y = map_data[loaded_map].enemy_spawn_location[2] + map_data[loaded_map].mget_shift[2] + dir.y
-  }
-  while not (cur.x == ending.x and cur.y == ending.y) do 
-    local north = {x=cur.x, y=cur.y-1}
-    local south = {x=cur.x, y=cur.y+1}
-    local west = {x=cur.x-1, y=cur.y}
-    local east = {x=cur.x+1, y=cur.y}
-    local state = false
-    local direction = nil
-    if dir.x == 1 then -- east 
-      state, direction = check_direction(east, {north, south}, path_tiles, path)
-    elseif dir.x == -1 then -- west
-      state, direction = check_direction(west, {north, south}, path_tiles, path)
-    elseif dir.y == 1 then -- south
-      state, direction = check_direction(south, {west, east}, path_tiles, path)
-    elseif dir.y == -1 then -- north
-      state, direction = check_direction(north, {west, east}, path_tiles, path)
-    end
-    assert(state, "Failed to find path at: "..cur.x..", "..cur.y.." in direction: "..dir.x..", "..dir.y.." end: "..ending.x..", "..ending.y)
-    if state then 
-      dir = {x=normalize(direction.x-cur.x), y=normalize(direction.y-cur.y)}
-      cur = {x=direction.x, y=direction.y}
-    else 
-    end
-  end
-  return path
-end
-
-function check_direction(direction, fail_directions, path_tiles, path)
-  if (direction == nil) return false, nil
-  local state, index = is_in_table(direction, path_tiles)
-  if state then
-    local tile = {
-      x = path_tiles[index].x - map_data[loaded_map].mget_shift[1],
-      y = path_tiles[index].y - map_data[loaded_map].mget_shift[2]
-    } 
-    add(path, tile)
-  else 
-    return check_direction(fail_directions[1], {fail_directions[2]}, path_tiles, path)
-  end
-  return true, direction
-end
-
-function place_tower(x, y)
-  -- check if there is a tower here
-  if (grid[y][x] == "tower") return false
-  -- check if player has the money
-  if (coins < tower_templates[shop_selector.pos + 1].cost) return false
-  -- spawn the tower
-  local tower_type = tower_templates[shop_selector.pos + 1].type 
-  if ((tower_type == "floor") ~= (grid[y][x] == "path")) return false 
-  add(towers, Tower:new(x, y, tower_templates[shop_selector.pos + 1], direction))
-  coins -= tower_templates[shop_selector.pos + 1].cost
-  grid[y][x] = "tower"
-  return true
-end
-
--- Text Helpers
-function print_with_outline(text, dx, dy, text_color, outline_color)
-  print(text, dx - 1, dy, outline_color)
-  print(text, dx + 1, dy, outline_color)
-  print(text, dx, dy-1, outline_color)
-  print(text, dx, dy+1, outline_color)
-  print(text, dx, dy, text_color)
-end
-
-function print_tower_cost(cost, dx, dy)
-  local color = 7
-  if (cost > coins) color = 8
-  print_with_outline("C"..cost, dx, dy, color, 0)
-end
-
--- Utility/Helper Functions
-function dist(posA, posB) 
-  local x = posA.x - posB.x
-  local y = posA.y - posB.y
-  return sqrt(x * x + y * y)
-end
-
-function clamp(val, min_val, max_val)
-  return min(max(min_val, val), max_val)
-end 
-
-function normalize(val)
-  return flr(clamp(val, -1, 1))
-end
-
-function lerp(start, last, rate)
-  return start + (last - start) * rate
-end
-
-function controls()
-  if btnp(⬆️) then return 0, -1
-  elseif btnp(⬇️) then return 0, 1
-  elseif btnp(⬅️) then return -1, 0
-  elseif btnp(➡️) then return 1, 0
-  end
-  return 0, 0
-end
-
-function increase_enemy_health(enemy_data)
-  return {
-    hp = enemy_data.hp + freeplay_stats.hp * freeplay_rounds,
-    step_delay = max(enemy_data.step_delay - freeplay_stats.speed * freeplay_rounds, freeplay_stats.min_step_delay),
-    sprite_index = enemy_data.sprite_index,
-    reward = enemy_data.reward,
-    damage = enemy_data.damage
-  }
-end
-
-function move_ui_selector(sel, dx, shift, offset, delta)
-  local inc = shift
-  if (dx < 0) inc = -shift
-  sel.pos = (sel.pos + inc) % #shop_ui_data.x
-  sel.x = shop_ui_data.x[sel.pos + offset]-delta
-end
-
-function is_in_table(val, table)
-  for i=1, #table do 
-    if (val.x == table[i].x and val.y == table[i].y) return true, i 
-  end
-  return false, -1
-end
-
-function get_flip_direction(direction)
-  return (direction[1] == -1), (direction[2] == -1)
-end
-
-function placable_tile_location(x, y, map_id)
-  local map_index = loaded_map
-  if (map_id ~= nil) map_index = map_id
-  local sprite_id = mget(x, y)
-  for i=1, #map_data[map_index].allowed_tiles do 
-    if (sprite_id == map_data[map_index].allowed_tiles[i]) return true
-  end
-  return false
-end
-
-function add_enemy_at_to_table(dx, dy, table, single_only)
-  for _, enemy in pairs(enemies) do
-    if (enemy.x == dx and enemy.y == dy) then
-      add(table, enemy)
-      if (single_only) return
-    end
-  end
-end
-
-function parse_direction(data, dir)
-  if dir[1] == 0 and dir[2] ~= 0 then 
-    return data[1]
-  elseif dir[1] ~= 0 and dir[2] == 0 then
-    return data[2]
-  end
-end
-
-function is_there_something_at(dx, dy, table)
-  for _, obj in pairs(table) do
-    if (obj.x == dx and obj.y == dy) return true 
-  end
-  return false
-end
-
-function refund_tower_at(dx, dy)
-  for _, tower in pairs(towers) do
-    if tower.x == dx and tower.y == dy then
-      grid[dy][dx] = "empty"
-      if (tower.type == "floor") grid[dy][dx] = "path"
-      coins += flr(tower.cost / 2) 
-      del(animators, tower.animator) 
-      del(towers, tower)
-    end
-  end
-end
-
-function parse_frontal_bounds(dx, dy, radius)
-  -- Default South
-  local fx, fy, flx, fly, ix, iy = -1, 1, 1, radius, 1, 1
-
-  if dx > 0 then -- east
-    fx, fy, flx, fly = 1, -1, radius, 1
-  elseif dx < 0 then -- west
-    fx, fy, flx, fly, ix = -1, -1, -radius, 1, -1
-  elseif dy < 0 then -- north
-    fx, fy, flx, fly, iy = -1, -1, 1, -radius, -1
-  end
-  return fx, fy, flx, fly, ix, iy
-end
-
--- particle spawing
-function raycast_spawn(dx, dy, range, direction, data)
-  for i=1, range do 
-    add(particles, Particle:new(dx + (i * direction[1]), dy + (i * direction[2]), false, Animator:new(data, false)))
-  end
-end
-
-function nova_spawn(dx, dy, radius, data)
-  for y=-radius, radius do
-    for x=-radius, radius do
-      if (x ~= 0 or y ~= 0) add(particles, Particle:new(dx + x, dy + y, false, Animator:new(data, false)))
-    end
-  end
-end
-
-function frontal_spawn(dx, dy, radius, direction, data)
-  local fx, fy, flx, fly, ix, iy = parse_frontal_bounds(direction[1], direction[2], radius)
-  for y=fy, fly, iy do
-    for x=fx, flx, ix do
-      if (x ~= 0 or y ~= 0) add(particles, Particle:new(dx + x, dy + y, false, Animator:new(data, false)))
-    end
-  end
-end
-
--- Draw Calls
-function game_draw_loop()
-  map(
-    map_data[loaded_map].data[1], 
-    map_data[loaded_map].data[2], 
-    map_data[loaded_map].data[3], 
-    map_data[loaded_map].data[4], 
-    map_data[loaded_map].data[5], 
-    map_data[loaded_map].data[6]
-  )
-  -- towers
-  foreach(towers, Tower.draw)
-  -- enemies
-  foreach(enemies, Enemy.draw)
-  -- particles
-  foreach(particles, Particle.draw)
-  -- shop
-  if (shop_enable) draw_shop_icons()
-  -- selector
-  if shop_enable then 
-    if option_enable then 
-      draw_selector(option_selector) 
-    else
-      draw_selector(shop_selector) 
-    end
-  else 
-    if not enemies_active and incoming_hint ~= nil then 
-      local dx = map_data[loaded_map].enemy_spawn_location[1]
-      local dy = map_data[loaded_map].enemy_spawn_location[2]
-      local dir = map_data[loaded_map].movement_direction
-      for i=1, #incoming_hint do 
-        spr(Animator.sprite_id(incoming_hint[i]), (dx + (i - 1) * dir[1])*8, (dy + (i - 1) * dir[2])*8) 
-      end
-    end
-    draw_selector(selector) 
-  end
-  -- UI
-  print_with_outline("scrap: "..coins, 0, 1, 7, 0)
-  print_with_outline("♥ "..player_health, 103, 1, 8, 0)
-  if shop_enable then
-    print_with_outline("game paused [ wave "..(wave_round+freeplay_rounds).." ]", 18, 16, 7, 0)
-    print_with_outline("start round", shop_ui_data.x[1]-6, 33, 7, 0)
-    print_with_outline("map menu", shop_ui_data.x[3]-6, 33, 7, 0)
-    local len = #tower_templates[shop_selector.pos + 1].name
-    print_with_outline(tower_templates[shop_selector.pos + 1].name, 128/2-(len*2), 108, 7, 0)
-    print_with_outline("❎ rotate | 🅾️ close shop", 1, 120, 7, 0)
-    draw_attack_tiles(tower_templates[shop_selector.pos + 1], 128/2 - 8, shop_ui_data.y[1] + 24)
-    draw_shop_cost()
-    draw_shop_dmg()
-  else 
-    local text = "❎ buy & place | 🅾️ open shop"
-    if (is_there_something_at(selector.x / 8, selector.y / 8, towers)) text = "❎ sell | 🅾️ open shop"
-    print_with_outline(text, 1, 122, 7, 0)
-  end
-end
-
-function draw_map_overview(map_id, xoffset, yoffset)
-  local mxshift = map_data[map_id].mget_shift[1]
-  local myshift = map_data[map_id].mget_shift[2]
-  for y=0, 15 do
-    for x=0, 15 do
-      local is_not_path = placable_tile_location(x + mxshift, y + myshift, map_id)
-      if is_not_path then 
-        pset(x + xoffset, y + yoffset, map_draw_data.other)
-      else 
-        pset(x + xoffset, y + yoffset, map_draw_data.path)
-      end
-    end
-  end
-end
-
-function draw_attack_tiles(tower_template, dx, dy)
-  if tower_template.type == "tack" then 
-    for y=-tower_template.radius, tower_template.radius do
-      for x=-tower_template.radius, tower_template.radius do
-        if (x ~= 0 or y ~= 0) spr(tile_display.attack, dx + x * 8, dy + y * 8)
-      end
-    end
-    spr(tower_template.sprite_data[1][2], dx, dy)
-  elseif tower_template.type == "rail" then 
-    local shift = (tower_template.radius + 1) / 2 - 1
-    for x=1, tower_template.radius do
-      if (x > 0) spr(tile_display.attack, dx + x * 8 - (shift * 8), dy)
-    end
-    spr(tower_template.sprite_data[1][2], dx - (shift * 8), dy)
-  elseif tower_template.type == "frontal" then 
-    for y=-1, 1 do
-      for x=1, tower_template.radius do 
-        spr(tile_display.attack, dx + x * 8, dy + y * 8)
-      end
-    end
-    spr(tower_template.sprite_data[1][2], dx, dy)
-  elseif tower_template.type == "floor" then 
-    spr(tile_display.attack, dx, dy)
-    spr(tower_template.sprite_data[1][2], dx, dy)
-  end
-end
-
-function draw_shop_icons()
-  local fx, fy = get_flip_direction(direction)
-  for i=1, #tower_templates do 
-    palt(0, false)
-    palt(14, false)
-    if (tower_templates[i].disable_icon_rotation) then 
-      spr(shop_ui_data.blank, shop_ui_data.x[i] - 20, shop_ui_data.y[1] - 20, 3, 3)
-      palt()
-      spr(parse_direction(tower_templates[i].icon_data, direction), shop_ui_data.x[i] - 16, shop_ui_data.y[1] - 16, 2, 2)
-    else
-      spr(parse_direction(shop_ui_data.background, direction), shop_ui_data.x[i] - 20, shop_ui_data.y[1] - 20, 3, 3, fx, fy)
-      palt()
-      spr(parse_direction(tower_templates[i].icon_data, direction), shop_ui_data.x[i] - 16, shop_ui_data.y[1] - 16, 2, 2, fx, fy)
-    end
-  end
-end
-
-function draw_shop_cost()
-  for i=1, #tower_templates do
-    print_tower_cost(tower_templates[i].cost, shop_ui_data.x[i] - 4, shop_ui_data.y[1] - 6)
-  end
-end
-
-function draw_shop_dmg()
-  for i=1, #tower_templates do
-    local type = tower_templates[i].type
-    if type == "tack" or type == "rail" then 
-      print_with_outline("D"..tower_templates[i].damage, shop_ui_data.x[i] - 4, shop_ui_data.y[1], 8, 0)
-    else 
-      print_with_outline("T"..tower_templates[i].damage, shop_ui_data.x[i] - 4, shop_ui_data.y[1], 12, 0)
-    end
-  end
-end
-
-function draw_selector(sel)
-  spr(sel.sprite_index, sel.x, sel.y, sel.size, sel.size)
 end
